fix: dev vs prod id type mismatch problem

+++ functions/studio-api/src/data/migrations/20221116185453-AddChoreListTables.ts
@@ -12,13 +12,13 @@ export default {
           {
             id: {
               primaryKey: true,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               unique: true,
               autoIncrement: true
             },
             assignee: {
               allowNull: false,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               primaryKey: true,
               references: {
                 model: 'users',
@@ -27,7 +27,7 @@ export default {
             },
             createdBy: {
               allowNull: false,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               primaryKey: true,
               references: {
                 model: 'users',
@@ -78,13 +78,13 @@ export default {
           {
             id: {
               primaryKey: true,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               unique: true,
               autoIncrement: true
             },
             choreChartId: {
               allowNull: false,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               onDelete: 'CASCADE',
               references: {
                 model: 'chore_charts',
@@ -122,12 +122,12 @@ export default {
           {
             id: {
               primaryKey: true,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               autoIncrement: true
             },
             choreChartId: {
               allowNull: false,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               onDelete: 'CASCADE',
               references: {
                 model: 'chore_charts',
@@ -137,7 +137,7 @@ export default {
             },
             choreId: {
               allowNull: false,
-              type: DataTypes.INTEGER,
+              type: DataTypes.BIGINT,
               onDelete: 'CASCADE',
               references: {
                 model: 'chores',
